// src/app/model/quiz.model.ts
import { createquiz, createquizdemo } from "./createquiz.model";

export class quiz {
  id: number = 0;
  quizname: string = '';
  questionName: createquiz[] = [];

  // quesitons: string[]=[ Option:string]:{
  //   option1: string="";
  //   option2: string;
  //   option3: string;
  //   option4: string;
  // }
}
export class quizdemo {
  id: number = 0;
  quizname: string = '';
  questionName: createquizdemo[] = [];

  // quesitons: string[]=[ Option:string]:{
  //   option1: string="";
  //   option2: string;
  //   option3: string;
  //   option4: string;
  // }
}